chore(main): rename bindContextMenu to bindMenus
