+++ src/main/util/bindMenus.js
